// FrontEnd/src/api/model/cadastroEventoDTOAreaAtuacao.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export type CadastroEventoDTOAreaAtuacao = typeof CadastroEventoDTOAreaAtuacao[keyof typeof CadastroEventoDTOAreaAtuacao];


export const CadastroEventoDTOAreaAtuacao = {
  ASSISTENCIA_SOCIAL: 'ASSISTENCIA_SOCIAL',
  DIREITOS_HUMANOS: 'DIREITOS_HUMANOS',
  MEIO_AMBIENTE: 'MEIO_AMBIENTE',
  SUSTENTABILIDADE: 'SUSTENTABILIDADE',
  EDUCACAO: 'EDUCACAO',
  CAPACITACAO: 'CAPACITACAO',
  SAUDE: 'SAUDE',
  BEM_ESTAR: 'BEM_ESTAR',
  CULTURA: 'CULTURA',
  LAZER: 'LAZER',
  ESPORTE: 'ESPORTE',
} as const;